Rename: FileUpload 관련 파일 이름에 Image 추가
